feat; split and update index and createServer function
